Fixed slow drive count
Previously, the stage was driven by a count of when B was pressed on gamepad 1. As the code runs several times a second, the count would go up much faster than intended. To combat this, I added a 'cooldown' effect with a timer. The count will only go up every 1.5 seconds between button presses. This should be the final change made to DriveSlow besides the actual speed.