Made Spin-up changes
Previously, due to the shooter needing to wait to reach full speed, the loops in teleop took 25 milliseconds longer than necessary. By adding a timer system, the 25 millisecond addon only lasts for the first 2 seconds of teleop. After these two seconds the shooter doesn't need to ramp up, so we can just run it at full speed and it no longer needs the extra 25 milliseconds to adjust to the new speed. This will result in a more streamlined code and driving experience.